Task 1 ChatBot/backend/requirements.txt
# FastAPI and server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# LangChain and LLM
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-qdrant>=0.1.0
langchain-huggingface>=0.0.1  # For updated HuggingFace embeddings

# Vector database
qdrant-client>=1.7.0
sentence-transformers>=2.6.1

# Environment and utilities
python-dotenv>=1.0.1
pydantic-settings>=2.1.0

# CORS support
python-jose[cryptography]>=3.3.0

# HTTP requests (for API Gateway)
requests>=2.32.5

# PDF processing for page number lookup
pdfplumber>=0.10.0

# Task scheduling
APScheduler>=3.10.0
python-dateutil>=2.8.2